Refactoring views.py.
