+++ api/views/__campus__.py
@@ -1,0 +1,12 @@
+from rest_framework import status
+from rest_framework.permissions import IsAuthenticatedOrReadOnly
+from rest_framework.viewsets import ModelViewSet
+from rest_framework.response import Response
+from rest_framework.decorators import api_view
+
+from api.serializers import CampusSerializer
+from api.models import Campus
+
+class CampusViewSet(ModelViewSet):
+    queryset = Campus.objects.all()
+    serializer_class = CampusSerializer
