+++ api/views/__init__.py
@@ -1,0 +1,2 @@
+from .__campus__ import CampusViewSet
+from .__course__ import CourseViewSet
